// src/app/components/auth/update-dtos/register-dtos/UpdateEo.dto.ts
import { AddressDTO } from "../../register-dtos/address.dto";

export interface UpdateEoDto {
    name: string | null | undefined;
    surname: string | null | undefined;
    phoneNumber: string | null | undefined;
    address: AddressDTO | null | undefined;
    photo: string | null | undefined;
}
  